Improve data trimming
The function that located the start and end point of the firing in raw data previously assumed that the datapoint with the largest value always was somewhere in the firing, but this sometimes isn't true, like in motors with powerful igniters or significant chuffing. In these cases, only data/stats from the chuff/igniter would be shown, which is obviously incorrect. This commit revises that function look at all ranges where values were above 0 to figure out which contains the firing before running the previous trimming algorithm.